# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LUMINA social network API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterInput {
  /**
     * @minLength 3
     * @maxLength 20
     */
  username: string;
  /** @minLength 1 */
  displayName: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginInput {
  /** Username or email */
  identifier: string;
  password: string;
}

export interface ProfileInput {
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  coverUrl?: string | null;
}

export interface PasswordInput {
  oldPassword: string;
  /** @minLength 8 */
  newPassword: string;
}

export interface User {
  id: number;
  username: string;
  displayName: string;
  email: string;
  bio: string;
  avatarUrl: string;
  coverUrl: string;
  website: string;
  location: string;
  verified: boolean;
  createdAt: string;
}

export interface AuthResult {
  token: string;
  user: User;
}

export interface UserSummary {
  id: number;
  username: string;
  displayName: string;
  avatarUrl: string;
  verified: boolean;
  bio?: string;
  isFollowing?: boolean;
  followerCount?: number;
}

export interface UserProfile {
  id: number;
  username: string;
  displayName: string;
  avatarUrl: string;
  coverUrl: string;
  bio: string;
  website: string;
  location: string;
  verified: boolean;
  createdAt: string;
  postCount: number;
  followerCount: number;
  followingCount: number;
  isFollowing: boolean;
}

export type PostInputPostType = typeof PostInputPostType[keyof typeof PostInputPostType];


export const PostInputPostType = {
  post: 'post',
  moment: 'moment',
} as const;

export interface PostInput {
  /** @minLength 1 */
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  /**
     * Second image for Moment posts
     * @nullable
     */
  imageUrl2?: string | null;
  postType?: PostInputPostType;
  /** @nullable */
  scheduledAt?: string | null;
}

export interface ReactionCount {
  reaction: string;
  count: number;
}

export interface Comment {
  id: number;
  postId: number;
  content: string;
  /** @nullable */
  parentCommentId?: number | null;
  createdAt: string;
  author: UserSummary;
  likeCount: number;
  liked: boolean;
}

export interface FeedPost {
  id: number;
  content: string;
  imageUrl: string;
  imageUrl2: string;
  postType: string;
  views: number;
  /** @nullable */
  scheduledAt?: string | null;
  createdAt: string;
  author: UserSummary;
  likeCount: number;
  commentCount: number;
  liked: boolean;
  saved: boolean;
  /** @nullable */
  myReaction?: string | null;
  topReactions: ReactionCount[];
  previewComments: Comment[];
}

export interface PostDetail {
  id: number;
  content: string;
  imageUrl: string;
  imageUrl2: string;
  postType: string;
  views: number;
  createdAt: string;
  author: UserSummary;
  likeCount: number;
  commentCount: number;
  liked: boolean;
  saved: boolean;
  /** @nullable */
  myReaction?: string | null;
  topReactions: ReactionCount[];
  comments: Comment[];
}

export interface LikeResult {
  liked: boolean;
  count: number;
  /** @nullable */
  reaction?: string | null;
}

export interface SaveResult {
  saved: boolean;
}

export interface CommentInput {
  /** @minLength 1 */
  content: string;
  /** @nullable */
  parentCommentId?: number | null;
}

export interface FollowResult {
  following: boolean;
  count: number;
}

export interface ReactionInput {
  /**
     * Emoji: ❤️ 🔥 😂 😮 😢 👏 or null for simple like
     * @nullable
     */
  reaction?: string | null;
}

export interface TrendingTag {
  tag: string;
  count: number;
}

export type StoryInputMediaType = typeof StoryInputMediaType[keyof typeof StoryInputMediaType];


export const StoryInputMediaType = {
  image: 'image',
  video: 'video',
  text: 'text',
} as const;

export interface StoryInput {
  /** @nullable */
  mediaUrl?: string | null;
  mediaType?: StoryInputMediaType;
  caption?: string;
  bgColor?: string;
  textContent?: string;
}

export interface Story {
  id: number;
  userId: number;
  mediaUrl: string;
  mediaType: string;
  caption: string;
  bgColor: string;
  textContent: string;
  views: number;
  expiresAt: string;
  createdAt: string;
}

export interface StoryGroup {
  user: UserSummary;
  stories: Story[];
  hasUnviewed: boolean;
}

export interface ViewResult {
  viewCount: number;
}

export type NotificationType = typeof NotificationType[keyof typeof NotificationType];


export const NotificationType = {
  like: 'like',
  comment: 'comment',
  follow: 'follow',
  mention: 'mention',
} as const;

export interface Notification {
  id: number;
  type: NotificationType;
  read: boolean;
  /** @nullable */
  postId?: number | null;
  /** @nullable */
  commentId?: number | null;
  createdAt: string;
  sender: UserSummary;
  /** @nullable */
  postPreview?: string | null;
}

export interface UnreadCount {
  count: number;
}

export type GetFeedParams = {
offset?: number;
limit?: number;
};

export type GetExplorePostsParams = {
offset?: number;
limit?: number;
/**
 * @nullable
 */
tag?: string | null;
};

export type GetSavedPostsParams = {
offset?: number;
limit?: number;
};

export type SearchUsersParams = {
q: string;
};

export type GetNotificationsParams = {
offset?: number;
limit?: number;
};