# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LUMINA social network API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update own profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Wrong password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── POSTS ───────────────────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get paginated notifications for current user
      security:
        - bearerAuth: []
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/unread-count:
    get:
      operationId: getUnreadCount
      tags: [notifications]
      summary: Get unread notification count
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnreadCount"
  /notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Marked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /notifications/read/{id}:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a single notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

# ── COMPONENTS ────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    # ── Auth schemas ────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [username, displayName, email, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 20
        displayName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginInput:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          description: Username or email
        password:
          type: string
    ProfileInput:
      type: object
      properties:
        displayName:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
    PasswordInput:
      type: object
      required: [oldPassword, newPassword]
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
    AuthResult:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # ── User schemas ─────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, username, displayName, email, bio, avatarUrl, coverUrl, website, location, verified, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        email:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        coverUrl:
          type: string
        website:
          type: string
        location:
          type: string
        verified:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserSummary:
      type: object
      required: [id, username, displayName, avatarUrl, verified]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
        verified:
          type: boolean
        bio:
          type: string
        isFollowing:
          type: boolean
        followerCount:
          type: integer
...
    Notification:
      type: object
      required: [id, type, read, createdAt, sender]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [like, comment, follow, mention]
        read:
          type: boolean
        postId:
          type: ["integer", "null"]
        commentId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        sender:
          $ref: "#/components/schemas/UserSummary"
        postPreview:
          type: ["string", "null"]
    UnreadCount:
      type: object
      required: [count]
      properties:
        count:
          type: integer