# lib/api-zod/src/generated/types/passwordInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LUMINA social network API
 * OpenAPI spec version: 0.1.0
 */

export interface PasswordInput {
  oldPassword: string;
  /** @minLength 8 */
  newPassword: string;
}